app/requirements.txt
ml-metadata==1.14.0; python_version < "3.12"
cyclonedx-python-lib
spdx-tools>=0.8.3
PyYAML>=6.0

